feat: popcorn total bytes transferred count query

+++ src/server/parser/parser.c
@@ -53,7 +53,7 @@ struct parser_event * parser_feed(struct parser *p, const uint8_t c) {
         const int when = state[i].when;
         if (when  <= 0xFF) {
             matched = (c == when);
-        } else if(when == ANY) {
+        } else if(when == (int) ANY) {
             matched = true;
         } else {
             matched = (type & when);
